+++ src/neuroconv/datainterfaces/ecephys/baserecordingextractorinterface.py
@@ -266,12 +266,10 @@ class BaseRecordingExtractorInterface(BaseExtractorInterface):
 
         return recording_extractor
 
-    def run_conversion(
+    def add_to_nwbfile(
         self,
-        nwbfile_path: Optional[FilePathType] = None,
-        nwbfile: Optional[NWBFile] = None,
+        nwbfile: NWBFile,
         metadata: Optional[dict] = None,
-        overwrite: bool = False,
         stub_test: bool = False,
         starting_time: Optional[float] = None,
         write_as: Literal["raw", "lfp", "processed"] = "raw",
@@ -286,18 +284,13 @@ class BaseRecordingExtractorInterface(BaseExtractorInterface):
 
         Parameters
         ----------
-        nwbfile_path : FilePathType
-            Path for where to write or load (if overwrite=False) the NWBFile.
-            If specified, this context will always write to this location.
-        nwbfile : NWBFile, optional
+        nwbfile : NWBFile
             NWBFile to which the recording information is to be added
         metadata : dict, optional
             metadata info for constructing the NWB file.
             Should be of the format::
 
                 metadata['Ecephys']['ElectricalSeries'] = dict(name=my_name, description=my_description)
-        overwrite: bool, default: False
-            Whether to overwrite the NWB file if one exists at the nwbfile_path.
         The default is False (append mode).
         starting_time : float, optional
             Sets the starting time of the ElectricalSeries to a manually set value.
@@ -337,20 +330,17 @@ class BaseRecordingExtractorInterface(BaseExtractorInterface):
                     Dictionary of keyword arguments to be passed directly to tqdm.
                     See https://github.com/tqdm/tqdm#parameters for options.
         """
-        from ...tools.spikeinterface import write_recording
+        from ...tools.spikeinterface import add_recording
 
         if stub_test or self.subset_channels is not None:
             recording = self.subset_recording(stub_test=stub_test)
         else:
             recording = self.recording_extractor
 
-        write_recording(
+        add_recording(
             recording=recording,
-            nwbfile_path=nwbfile_path,
             nwbfile=nwbfile,
             metadata=metadata,
-            overwrite=overwrite,
-            verbose=self.verbose,
             starting_time=starting_time,
             write_as=write_as,
             write_electrical_series=write_electrical_series,