+++ src/neuroconv/datainterfaces/icephys/baseicephysinterface.py
@@ -82,16 +82,12 @@ class BaseIcephysInterface(BaseExtractorInterface):
     def align_by_interpolation(self, unaligned_timestamps: np.ndarray, aligned_timestamps: np.ndarray):
         raise NotImplementedError("Icephys interfaces do not yet support timestamps.")
 
-    def run_conversion(
+    def add_to_nwbfile(
         self,
-        nwbfile: NWBFile = None,
-        nwbfile_path: Optional[FilePathType] = None,
+        nwbfile: NWBFile,
         metadata: dict = None,
-        overwrite: bool = False,
         icephys_experiment_type: str = "voltage_clamp",
         skip_electrodes: Tuple[int] = (),
-        # TODO: to be removed
-        save_path: Optional[FilePathType] = None,  # pragma: no cover
     ):
         """
         Primary function for converting raw (unprocessed) intracellular data to the NWB standard.
@@ -100,46 +96,17 @@ class BaseIcephysInterface(BaseExtractorInterface):
         ----------
         nwbfile : NWBFile
             nwb file to which the recording information is to be added
-        nwbfile_path : FilePathType
-            Path for where to write or load (if overwrite=False) the NWBFile.
-            If specified, the context will always write to this location.
         metadata : dict, optional
             metadata info for constructing the nwb file (optional).
-        overwrite : bool, default: False
-            Whether to overwrite the NWB file if one exists at the nwbfile_path.
         icephys_experiment_type : {'voltage_clamp', 'current_clamp', 'izero'}
             Type of icephys recording.
         skip_electrodes : tuple, optional
             Electrode IDs to skip. Defaults to ().
-        save_path: string, optional
         """
-        from ...tools.neo import write_neo_to_nwb
+        from ...tools.neo import add_neo_to_nwb
 
         if nwbfile is None:
             nwbfile = make_nwbfile_from_metadata(metadata)
-
-        # TODO on or after August 1st, 2022, remove argument and deprecation warnings
-        if save_path is not None:  # pragma: no cover
-            will_be_removed_str = "will be removed on or after October 1st, 2022. Please use 'nwbfile_path' instead."
-            if nwbfile_path is not None:
-                if save_path == nwbfile_path:
-                    warn(
-                        "Passed both 'save_path' and 'nwbfile_path', but both are equivalent! "
-                        f"'save_path' {will_be_removed_str}",
-                        DeprecationWarning,
-                    )
-                else:
-                    warn(
-                        "Passed both 'save_path' and 'nwbfile_path' - using only the 'nwbfile_path'! "
-                        f"'save_path' {will_be_removed_str}",
-                        DeprecationWarning,
-                    )
-            else:
-                warn(
-                    f"The keyword argument 'save_path' to 'spikeinterface.write_recording' {will_be_removed_str}",
-                    DeprecationWarning,
-                )
-                nwbfile_path = save_path
 
         if (
             HAVE_NDX_DANDI_ICEPHYS
@@ -149,16 +116,11 @@ class BaseIcephysInterface(BaseExtractorInterface):
             nwbfile.add_lab_meta_data(DandiIcephysMetadata(**metadata["ndx-dandi-icephys"]))
 
         for i, reader in enumerate(self.readers_list):
-            write_neo_to_nwb(
+            add_neo_to_nwb(
                 neo_reader=reader,
                 nwbfile=nwbfile,
                 metadata=metadata,
-                overwrite=overwrite if i == 0 else False,
                 icephys_experiment_type=metadata["Icephys"]["Sessions"][i]["icephys_experiment_type"],
                 stimulus_type=metadata["Icephys"]["Sessions"][i]["stimulus_type"],
                 skip_electrodes=skip_electrodes,
             )
-
-        if save_path:
-            with NWBHDF5IO(save_path, "w") as io:
-                io.write(nwbfile)